Add localstack to fake s3
Previously we've needed to use a real s3 bucket
(digitalmarketplace-dev-uploads) for local instances of dmrunner to be
able to do s3 things.

This commit (in combination with alphagov/digitalmarketplace-utils#580
and alphagov/digitalmarketplace-supplier-frontend#1276) adds a
localstack instance providing s3 and hooks it up to the Digital
Marketplace apps.

Note that we have to set AWS_ACCESS_KEY_ID and AWS_SECRET_ACCESS_KEY to
`test`, because localstack uses these values when checking signatures
for presigned s3 urls [[1]].

[1]: https://github.com/localstack/localstack/blob/68d02ebf2ec05f1959bb6601aee5e1999e057a17/localstack/constants.py#L120-L122